+++ src/app/app.routes.ts
@@ -81,6 +81,7 @@ import { ModalDialogDemoComponent } from './modal-dialog/modal-dialog.demo';
 import { PopDownDemoComponent } from './popdown/popdown.demo';
 import { PopupMenuDemoComponent } from './popupmenu/popupmenu.demo';
 import { ProgressDemoComponent } from './progress/progress.demo';
+import { RadarDemoComponent } from './radar/radar.demo';
 import { RadioButtonDemoComponent } from './radiobutton/radiobutton.demo';
 import { SearchFieldDemoComponent } from './searchfield/searchfield.demo';
 import { SliderDemoComponent } from './slider/slider.demo';
@@ -210,6 +211,7 @@ export const routes: Routes = [
   { path: 'popdown',                                component: PopDownDemoComponent},
   { path: 'popupmenu',                              component: PopupMenuDemoComponent },
   { path: 'progress',                               component: ProgressDemoComponent },
+  { path: 'radar',                                  component: RadarDemoComponent},
   { path: 'radiobutton',                            component: RadioButtonDemoComponent},
   { path: 'searchfield',                            component: SearchFieldDemoComponent },
   { path: 'slider',                                 component: SliderDemoComponent },