+++ src/soho/radar/index.ts
@@ -1,0 +1,2 @@
+export * from './soho-radar.component';
+export * from './soho-radar.module';